@@ -18,16 +18,16 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 0.0.1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  auto_size_text: ^2.1.0
-  cupertino_icons: ^1.0.0
-  google_fonts: ^1.1.1
-  modal_bottom_sheet: ^2.0.0-nullsafety.1
-  provider: ^4.3.2+2
+  auto_size_text: ^3.0.0-nullsafety.0
+  cupertino_icons: ^1.0.2
+  google_fonts: ^2.0.0
+  modal_bottom_sheet: ^2.0.0
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
