@@ -1,9 +1,9 @@
 class Choice {
-  final String text;
+  final String? text;
 
   /// Default value is null. Otherwise, value is true.
-  final bool isAnswer;
-  final int id;
+  final bool? isAnswer;
+  final int? id;
 
   Choice({this.text, this.isAnswer, this.id});
 }
