@@ -24,7 +24,7 @@ class MyApp extends StatelessWidget {
             textTheme: TextTheme(
               headline6: Theme.of(context)
                   .textTheme
-                  .headline6
+                  .headline6!
                   .copyWith(color: Colors.black),
             ),
             iconTheme: IconThemeData(color: Colors.black),
